+++ Controls/_popup/Manager/Popup.ts
@@ -11,18 +11,13 @@ import {delay as runDelayed} from 'Types/function';
 
 import * as template from 'wml!Controls/_popup/Manager/Popup';
 import * as PopupContent from 'wml!Controls/_popup/Manager/PopupContent';
+import {IPrefetchData, waitPrefetchData} from 'Controls/_popup/utils/Preload';
 
 const RESIZE_DELAY = 10;
 
 interface IPopupControlOptions extends IPopupOptions, IControlOptions {
     _prefetchPromise?: Promise<IPrefetchPromises>; // TODO: Compatible предзагрузка
     prefetchData?: Record<string, unknown>; // TODO: Compatible предзагрузка
-}
-interface IPrefetchPromises {  // TODO: Compatible предзагрузка
-    [key: string]: Promise<unknown>;
-}
-interface IPrefetchData {  // TODO: Compatible предзагрузка
-    [key: string]: unknown;
 }
 /**
  * Control Popup
@@ -67,20 +62,21 @@ class Popup extends Control<IPopupControlOptions> {
 
         this._controlResizeHandler = debounce(this._controlResizeHandler.bind(this), RESIZE_DELAY, true);
 
-        if (options._prefetchPromise) {
-            const prefetchPromise = this._preparePrefetchData(options._prefetchPromise).then((data: IPrefetchData) => {
+        if (options._prefetchData) {
+            this._prefetchData = options._prefetchData;
+            this._isPrefetchDataMode = true;
+        } else if (options._prefetchPromise) {
+            waitPrefetchData(options._prefetchPromise).then((data: IPrefetchData) => {
                 this._prefetchData = data;
+                this._isPrefetchDataMode = true;
             });
-            if (options.initializingWay === 'remote') {
-                return prefetchPromise;
-            }
         } else if (options.prefetchData) {
             // Если с прикладной стороны сами получили предзагруженные данные
             this._isPrefetchDataMode = true;
         }
     }
 
-    //TODO: https://online.sbis.ru/opendoc.html?guid=728a9f94-c360-40b1-848c-e2a0f8fd6d17
+    // TODO: https://online.sbis.ru/opendoc.html?guid=728a9f94-c360-40b1-848c-e2a0f8fd6d17
     private _getCompatibleTemplateName(options: IPopupOptions): string {
         if (options.isCompoundTemplate) {
             return options.templateOptions.template;
@@ -122,10 +118,14 @@ class Popup extends Control<IPopupControlOptions> {
 
     protected _beforeUpdate(options: IPopupControlOptions): void {
         this._stringTemplate = typeof options.template === 'string';
-        if (options._prefetchPromise !== this._options._prefetchPromise) {
+        if (options._prefetchData !== this._options._prefetchData && options._prefetchData) {
+            this._prefetchData = options._prefetchData;
+            this._isPrefetchDataMode = true;
+        } else if (options._prefetchPromise !== this._options._prefetchPromise) {
             if (options._prefetchPromise) {
-                this._preparePrefetchData(options._prefetchPromise).then((data: IPrefetchData) => {
+                waitPrefetchData(options._prefetchPromise).then((data: IPrefetchData) => {
                     this._prefetchData = data;
+                    this._isPrefetchDataMode = true;
                 });
             } else {
                 this._prefetchData = null;
@@ -143,24 +143,6 @@ class Popup extends Control<IPopupControlOptions> {
             this._startResizeRegister();
             this._checkResizeObserver();
         }
-    }
-
-    private _preparePrefetchData(prefetchPromise: Promise<IPrefetchPromises>): Promise<IPrefetchData> {
-        this._isPrefetchDataMode = true;
-        return prefetchPromise.then((prefetchPromiseData: IPrefetchPromises) => {
-            const promiseArray = Object.values(prefetchPromiseData);
-            return Promise.allSettled(promiseArray).then(
-                (dataArray: Array<{status: string, value?: unknown, reason?: unknown}>) => {
-                    const keys = Object.keys(prefetchPromiseData);
-                    const data = {};
-                    let i = 0;
-                    for (const key of keys) {
-                        const promiseResult = dataArray[i++];
-                        data[key] = promiseResult.value || promiseResult.reason;
-                    }
-                    return data;
-                });
-        });
     }
 
     private _startResizeRegister(event?: SyntheticEvent): void {
